Refactor 'reserve' and 'stableCoinFactory' tests for code clarity
The export function wrapper for the 'reserve' and 'stableCoinFactory' test suites has been removed to improve clarity. Key functionalities of the tests remain unchanged.
