+++ contracts/test/deployFactory.ts
@@ -2,11 +2,9 @@ import '@hashgraph/hardhat-hethers'
 import { expect } from 'chai'
 import { deployFactory } from '../scripts/deployFactory'
 
-export const deployAFactory = () => {
-    describe('Deploy a Factory', function () {
-        it('Deploys the factory', async function () {
-            await deployFactory()
-            expect(true).to.equals(true)
-        })
+describe('Deploy a Factory', function () {
+    it('Deploys the factory', async function () {
+        await deployFactory()
+        expect(true).to.equals(true)
     })
-}
+})
